requirements.txt
alive-progress==1.6.1
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
ciso8601==2.1.3
click==7.1.2
clickhouse-driver==0.2.0
colorama==0.4.4
colorlog==4.6.2
commonmark==0.9.1
coverage==5.5
DateTime==4.3
future==0.18.2
gprof2dot==2019.11.30
idna==2.10
importlib-metadata==3.7.2
iniconfig==1.1.1
kafka-python==2.0.2
llvmlite==0.35.0
msgpack==1.0.2
numba==0.52.0
numpy==1.20.2
orjson==3.5.2
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyArango==1.3.4
Pygments==2.8.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.3
pytest-clarity==0.3.0a0
python-arango==6.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-toolbelt==0.9.1
rich==10.1.0
six==1.15.0
snakeviz==2.1.0
termcolor==1.1.0
toml==0.10.2
tornado==6.1
tqdm==4.60.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
zipp==3.4.1
zope.interface==5.4.0
